fix nms config typing

+++ yolo/config/config.py
@@ -102,8 +102,8 @@ class EMAConfig:
 
 @dataclass
 class NMSConfig:
-    min_confidence: int
-    min_iou: int
+    min_confidence: float
+    min_iou: float
 
 
 @dataclass
